### 2020-09-01 #2625 @whedon
[PRE REVIEW]: nap: A molecular dynamics package with parameter-optimization programs for classical and machine-learning potentials
**Submitting author:** @ryokbys (<a href="http://orcid.org/0000-0001-8244-5844">Ryo KOBAYASHI</a>)
**Repository:** <a href="https://github.com/ryokbys/nap" target ="_blank">https://github.com/ryokbys/nap</a>
**Version:** v0.9.2
