+++ doc/intro.rst
@@ -8,17 +8,17 @@ And *pmd* is a part of the **Nagoya atomistic-simulation package (nap)**.
 
 The main features of *pmd* are the following:
 
-* several interatomic potentials for solid state systems are available;
-* **Deep neural-network (DNN) interatomic potential**;
-* **QEq** or **variable charge** Coulombic potential;
-* parallel computation using spatial decomposition technique;
-* efficient searching of neighbor atoms using linked-list cell method;
-* structure relaxation using simple velocity damping or **FIRE** algorithm;
-* thermostats: Berendsen and Langevin;
-* barostat: Berendsen;
-* **variable-timestep** MD for high-energy ion-bombardment simulation;
-* **non-equilibrium MD (NEMD)** for heat flux simulation;
-* **two-temperature model MD (TTM-MD)** for laser-ablation simulation.
+  * several interatomic potentials for solid state systems are available;
+  * **Deep neural-network (DNN) interatomic potential**;
+  * **QEq** or **variable charge** Coulombic potential;
+  * parallel computation using spatial decomposition technique;
+  * efficient searching of neighbor atoms using linked-list cell method;
+  * structure relaxation using simple velocity damping or **FIRE** algorithm;
+  * thermostats: Berendsen and Langevin;
+  * barostat: Berendsen;
+  * **variable-timestep** MD for high-energy ion-bombardment simulation;
+  * **non-equilibrium MD (NEMD)** for heat flux simulation;
+  * **two-temperature model MD (TTM-MD)** for laser-ablation simulation.
 
 Since this program has been developed for the purpose of personal research tool,
 there are not so many functionalities. 
@@ -34,18 +34,6 @@ but I am afraid that I might not be able to respond all the reports or questions
 
 Requirements
 ====================
-*pmd* can be executed in,
+*pmd* can be executed in Unix/Linux, macOS X, and Windows with using a Fortran compiler and an MPI library.
 
-* Linux;
-* MacOS X;
-* Windows (Cygwin).
-
-You also need,
-
-* Fortran compiler;
-* MPI library;
-
-installed in the system you want to run ``pmd``.
-
-For some analysis tools,
-you may also need *Python 3.* and some python-utilities such as *numpy*, *scipy* and *pandas*.
+For some analysis tools written in Python language, you may also need *Python 3.* and some python-utilities such as *numpy*, *scipy*, *pandas*, *docopt* and *ASE*.
