### 2020-09-08 #2625 @usethedata
Possible, but I'll be a bit slower than usual.  This is end of fiscal year time in the midst of a reorganization, so my work load is a bit crazy.
