+++ doc/conf.py
@@ -1,279 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# nap_manual documentation build configuration file, created by
-# sphinx-quickstart on Wed Jan 29 22:34:45 2014.
-#
-# This file is execfile()d with the current directory set to its containing dir.
-#
-# Note that not all possible configuration values are present in this
-# autogenerated file.
-#
-# All configuration values have a default; values that are commented out
-# serve to show the default.
-
-import sys, os
-import sphinx_rtd_theme
-#import guzzle_sphinx_theme
-from recommonmark.parser import CommonMarkParser
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#sys.path.insert(0, os.path.abspath('.'))
-sys.path.append(os.path.abspath('_themes'))
-
-# -- General configuration -----------------------------------------------------
-
-# If your documentation needs a minimal Sphinx version, state it here.
-#needs_sphinx = '1.0'
-
-# Add any Sphinx extension module names here, as strings. They can be extensions
-# coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['sphinx.ext.mathjax','recommonmark']
-#extensions.append('guzzle_sphinx_theme')
-pngmath_latex_preamble = r"""
-\usepackage{eulervm}
-\usepackage{amsmath,amssymb}
-\newcommand{\degree}{$^{\circ}$}
-\newcommand{\temp}{${}^\circ$C}
-\newcommand{\rmd}{{\rm d}}
-\newcommand{\rme}{{\rm e}}
-"""
-
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
-
-# The suffix of source filenames.
-#source_suffix = '.rst'
-# source_suffix = ['.rst','.md']
-source_suffix = {
-    '.rst': 'restructuredtext',
-    '.md': 'markdown',
-}
-# source_parsers = {
-#     '.md' : 'recommonmark.parser.CommonMarkParser',
-# }
-
-# The encoding of source files.
-#source_encoding = 'utf-8-sig'
-
-# The master toctree document.
-master_doc = 'index'
-
-# General information about the project.
-project = u'nap documentation'
-copyright = u'2014-2020, Ryo KOBAYASHI'
-
-# The version info for the project you're documenting, acts as replacement for
-# |version| and |release|, also used in various other places throughout the
-# built documents.
-#
-# The short X.Y version.
-version = 'rev200722'
-# The full version, including alpha/beta/rc tags.
-release = 'rev200722'
-
-# The language for content autogenerated by Sphinx. Refer to documentation
-# for a list of supported languages.
-language = 'en'
-
-# There are two options for replacing |today|: either, you set today to some
-# non-false value, then it is used:
-#today = ''
-# Else, today_fmt is used as the format for a strftime call.
-#today_fmt = '%B %d, %Y'
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-exclude_patterns = ['_build']
-
-# The reST default role (used for this markup: `text`) to use for all documents.
-#default_role = None
-
-# If true, '()' will be appended to :func: etc. cross-reference text.
-#add_function_parentheses = True
-
-# If true, the current module name will be prepended to all description
-# unit titles (such as .. function::).
-#add_module_names = True
-
-# If true, sectionauthor and moduleauthor directives will be shown in the
-# output. They are ignored by default.
-#show_authors = False
-
-# The name of the Pygments (syntax highlighting) style to use.
[…]
