+++ mkdocs/upload.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+mkdocs build
+base=site
+rsync -avz ${base}/* nitweb:httpdocs/contents/nap_docs/
