+++ mkdocs/docs/analysis.md
@@ -1,0 +1,189 @@
+# How to analyze simulation results
+
+
+------------------------------------------------------------------------
+
+## Plot evolution of energies
+
+
+Energies are basic values in MD. If you perform *NVE* simulation in
+which no damping, temperature control, nor external forces, the total
+energy has to be conserved whereas kinetic and potential energies
+fluctuate. This is often used to check the validity of forces on atoms
+when programming new potentials.
+
+When one performs structure relaxation by damping velocities of atoms,
+as MD steps increase the kinetic energy has to decrease to zero.
+
+When one performs constant temperture simulation, kinetic energy should
+be almost constant during the simulation.
+
+Since total, kinetic, and potential energies are written in `out.erg`
+file, users can plot energy evolution using *gnuplot* command as, :
+
+    $ gnuplot
+    gnuplot> plot 'out.erg' us 1:3 w l t 'total', 'out.erg' us 1:4 w l t 'kinetic', 'out.erg' us 1:5 w l t 'potential'
+
+Or copy `util/gp.erg` script to the working directory and, :
+
+    gnuplot> load 'gp.erg'
+
+------------------------------------------------------------------------
+
+## Visualize atom configuration
+
+### Convert from pmd-format files
+
+There is a conversion program that changes from pmd format to
+visualization software format. When visualizing the atom configuration
+by using [Ovito](https://www.ovito.org) , first convert the *pmd*-format
+files, `pmd_###` , to *LAMMPS-dump* format files by doing the following,
+
+    $ /path/to/nap/nappy/napsys.py convert --specorder=A,B,C pmd_#### dump_####
+
+where `--specorder=A,B,C` specifies the order of species used in the
+pmd-format file. Then you get `dump_####` file where `####` should be a
+4-digit number. The [Ovito](https://www.ovito.org) can open the
+LAMMPS-dump format file.
+
+If there are sequential `pmd_####` files, one can convert those files by
+using bash for-statement as :
+
+    $ for f in pmd_[0-9]*; do /path/to/nap/nappy/napsys.py convert \
+        --specorder=A,B,C $f `echo $f | sed 's/pmd/dump/'`; done
+
+### Write LAMMPS-dump file directly
+
+If you set `flag_out_pmd` as `2`, the *pmd* program writes atomic
+configurations in *LAMMPS-dump* format with file names `dump_####`. So
+you can visualize those files without any conversion using *Ovito* or
+some other programs that can visualize *LAMMPS-dump* file.
+
+------------------------------------------------------------------------
+
+## Radial distribution function (RDF)
+
+RDF is commonly used to investigate structure of the system, e.g., solid vs liquid, or crystal vs amorphous.
+
+To get the RDF,
+
+    $ python /path/to/nap/nappy/rdf.py [options] dump_0*
+
+then, you get averaged RDF over atoms in `out.rdf`.
+
+Given atom configuration files, `dump_####`, are read and average over
+atoms in those files are taken.
+
+Options are shown below,
+
+    Options:
+      -h, --help  Show this help message and exit.
+      -d DR       Width of the bin. [default: 0.1]
+      -r RMAX     Cutoff radius of radial distribution. [default: 5.0]
+      --gsmear=SIGMA
+                  Width of Gaussian smearing, zero means no smearing. [default: 0]
+      -o OUT      Output file name. [default: out.rdf]
+      --specorder=SPECORDER
+                  Order of species separated by comma, like, --specorder=W,H. [default: None]
+      --skip=NSKIP 
+                  Skip first NSKIP steps from the statistics. [default: 0]
+      --no-average
+                  Not to take average over files.
+      --no-normalize
+                  Not to normalize by the density.
+      --plot      Plot figures. [default: False]
+
[…]
