+++ README.md
@@ -34,7 +34,7 @@ $ ../../pmd/pmd
 ```
 
 If it works, you can use this program.
-For details, please see the [documentation](http://ryokbys.web.nitech.ac.jp/contents/pmd_usage) or ask me via e-mail (kobayashi.ryo[at]nitech.ac.jp).
+For details, please see the [documentation](http://ryokbys.web.nitech.ac.jp/contents/nap_docs) or ask me via e-mail (kobayashi.ryo[at]nitech.ac.jp).
 
 # Acknowledgements
 This program was supported in part by ["Materials research by Information Integration" Initiative (MI2I)](http://www.nims.go.jp/MII-I/) project of the Support Program for Starting Up Innovation Hub from Japan Science and Technology Agency (JST).
