### 2020-09-10 57461d4
fixed the link of logo.
