### 2020-09-10 24fecd0
fixed the link of logo.
