### 2020-10-23 da1af85
modified force_LJ not to avoid double counting.
