### 2020-10-23 f332158
refactored force_LJ.
