### 2020-10-23 97326fd
modified some force_xxx not to avoid double count.
