### 2020-10-23 #2768 @whedon
[REVIEW]: nap: A molecular dynamics package with parameter-optimization programs for classical and machine-learning potentials
**Submitting author:** @ryokbys (<a href="http://orcid.org/0000-0001-8244-5844">Ryo KOBAYASHI</a>)
**Repository:** <a href="https://github.com/ryokbys/nap" target ="_blank">https://github.com/ryokbys/nap</a>
**Version:** 0.10.3
