+++ pmd/mod_pmdio.F90
@@ -1,16 +1,18 @@
 module pmdio
 !-----------------------------------------------------------------------
-!                     Last modified: <2020-06-16 18:24:08 Ryo KOBAYASHI>
+!                     Last modified: <2020-11-06 22:24:06 Ryo KOBAYASHI>
 !-----------------------------------------------------------------------
   implicit none
   save
 
   character(len=20),parameter:: cinpmd='in.pmd'
 
-  integer:: ntot0,ntot
+  integer:: ntot0,ntot,naux
 !.....data of total system
   real(8),allocatable:: rtot(:,:),vtot(:,:),stot(:,:,:),epitot(:) &
-       ,ekitot(:,:,:),tagtot(:),atot(:,:),chgtot(:),chitot(:),teitot(:)
+       ,ekitot(:,:,:),tagtot(:),atot(:,:)
+  real(8),allocatable:: chgtot(:),chitot(:),teitot(:),clrtot(:)
+  real(8),allocatable:: auxtot(:,:)
   real(8):: hunit,h(3,3,0:1)
 
 !.....max. num. of atoms in a node
@@ -292,16 +294,16 @@ contains
          xlo_bound,xhi_bound,ylo_bound,yhi_bound, &
          zlo_bound,zhi_bound,st(3,3)
 !!$    integer,external:: itotOf
-    real(8),allocatable,save:: rlmp(:,:)
+    real(8),allocatable,save:: rlmp(:,:),vlmp(:,:)
     character(len=3):: csp
     character(len=6):: caux
 
     real(8),parameter:: tiny = 1d-14
 
-    if( .not. allocated(rlmp) ) allocate(rlmp(3,ntot))
+    if( .not. allocated(rlmp) ) allocate(rlmp(3,ntot),vlmp(3,ntot))
     if( size(rlmp).ne.3*ntot ) then
-      deallocate(rlmp)
-      allocate(rlmp(3,ntot))
+      deallocate(rlmp,vlmp)
+      allocate(rlmp(3,ntot),vlmp(3,ntot))
     endif
 
     open(ionum,file=trim(cfname),status='replace')
@@ -310,7 +312,8 @@ contains
     write(ionum,'(a)') 'ITEM: NUMBER OF ATOMS'
     write(ionum,'(i10)') ntot
     write(ionum,'(a)') 'ITEM: BOX BOUNDS xy xz yz'
-    call pmd2lammps(h,ntot,rtot,rlmp,xlo,xhi,ylo,yhi,zlo,zhi,xy,xz,yz)
+    call pmd2lammps(h,ntot,rtot,rlmp,vtot,vlmp &
+         ,xlo,xhi,ylo,yhi,zlo,zhi,xy,xz,yz)
     xlo_bound = xlo +min(0d0, xy, xz, xy+xz)
     xhi_bound = xhi +max(0d0, xy, xz, xy+xz)
     ylo_bound = ylo +min(0d0, yz)
@@ -354,6 +357,12 @@ contains
             write(ionum,'(es11.3)',advance='no') eki
           else if( trim(caux).eq.'epot'.or.trim(caux).eq.'epi' ) then
             write(ionum,'(es11.3)',advance='no') epi
+          else if( trim(caux).eq.'vx' ) then
+            write(ionum,'(es11.3)',advance='no') vlmp(1,i)
+          else if( trim(caux).eq.'vy' ) then
+            write(ionum,'(es11.3)',advance='no') vlmp(2,i)
+          else if( trim(caux).eq.'vz' ) then
+            write(ionum,'(es11.3)',advance='no') vlmp(3,i)
[…]
