+++ pmd/pmd_core.F90
@@ -1,10 +1,10 @@
 !-----------------------------------------------------------------------
-!                     Last-modified: <2020-06-05 20:07:26 Ryo KOBAYASHI>
+!                     Last-modified: <2020-11-06 21:08:29 Ryo KOBAYASHI>
 !-----------------------------------------------------------------------
 ! Core subroutines/functions needed for pmd.
 !-----------------------------------------------------------------------
 subroutine pmd_core(hunit,h,ntot0,tagtot,rtot,vtot,atot,stot &
-     ,ekitot,epitot,chgtot,chitot,teitot,maxstp,nerg,npmd &
+     ,ekitot,epitot,chgtot,chitot,teitot,clrtot,maxstp,nerg,npmd &
      ,myid_md,mpi_md_world,nodes_md,nx,ny,nz,specorder &
      ,am,dt,vardt_len,ciofmt,ifpmd,rc,rbuf,rc1nn,ifdmp,dmp &
      ,minstp,tinit,tfin,ctctl,ttgt,trlx,ltdst,ntdst,nrmtrans,cpctl &
@@ -29,6 +29,8 @@ subroutine pmd_core(hunit,h,ntot0,tagtot,rtot,vtot,atot,stot &
   use structure,only: cna,acna
   use deform,only: init_deform, apply_deform
   use util,only: itotOf, ifmvOf
+  use extforce,only: lextfrc,rm_trans_extfrc,add_extfrc
+  use clrchg,only: lclrchg,clrchg_force
   implicit none
   include "mpif.h"
   include "./params_unit.h"
@@ -44,7 +46,8 @@ subroutine pmd_core(hunit,h,ntot0,tagtot,rtot,vtot,atot,stot &
        ,atot(3,ntot0),stot(3,3,ntot0),ekitot(3,3,ntot0) &
        ,epitot(ntot0),dt,rc,rbuf,rc1nn,h(3,3,0:1),stnsr(3,3) &
        ,fmv(3,0:9),epot,ekin,am(nspmax),stgt(3,3),ptgt,pini &
-       ,pfin,ttgt(9),chgtot(ntot0),chitot(ntot0),teitot(ntot0)
+       ,pfin,ttgt(9),chgtot(ntot0),chitot(ntot0),teitot(ntot0) &
+       ,clrtot(ntot0)
   character,intent(in):: ciofmt*6, cpctl*20, ctctl*20 &
        ,boundary*3
   character(len=3),intent(in):: specorder(nspmax) 
@@ -167,7 +170,7 @@ subroutine pmd_core(hunit,h,ntot0,tagtot,rtot,vtot,atot,stot &
   endif
 !.....perform space decomposition after reading atomic configuration
   call space_decomp(hunit,h,ntot0,tagtot,rtot,vtot,chgtot,chitot,teitot &
-       ,myid_md,mpi_md_world,nx,ny,nz,nxyz,rc,rbuf,iprint)
+       ,clrtot,myid_md,mpi_md_world,nx,ny,nz,nxyz,rc,rbuf,iprint)
 !.....Some conversions
   do i=1,natm
     ra(1:3,i)= ra(1:3,i) -sorg(1:3)
@@ -273,8 +276,12 @@ subroutine pmd_core(hunit,h,ntot0,tagtot,rtot,vtot,atot,stot &
   elseif( abs(tinit).le.1d-5 ) then
     va(1:3,1:natm)= 0d0
   endif
-  if( nrmtrans.ge.0 ) call rm_trans_motion(natm,tag,va,nspmax,am &
-       ,mpi_md_world,myid_md,iprint)
+  if( lextfrc ) then  ! special treatment for translational momentum
[…]
