### 2020-11-06 f9603ee
fixed a bug on read_dump in napsys.py
