### 2020-11-06 #2768 @usethedata
@ryokbys -- while I'm looking at this, I agree with Whedon's suggestion to add `, doi = 10.1093/oso/9780198803195.001.0001` to the bibliographic entry for the Allen and Tildesley book.  No hurries, but thought I'd tag it to your attention.
