+++ JOSS_paper/paper.bib
@@ -119,6 +119,7 @@
   year      =  2017,
   language  = "en",
   isbn      = "9780198803201"
+  doi       = "10.1093/oso/9780198803195.001.0001"
 }
 
 @ARTICLE{Artrith2016-mu,
