+++ JOSS_paper/paper.bib
@@ -118,7 +118,7 @@
   month     =  aug,
   year      =  2017,
   language  = "en",
-  isbn      = "9780198803201"
+  isbn      = "9780198803201",
   doi       = "10.1093/oso/9780198803195.001.0001"
 }
 
