### 2020-11-07 #2768 @ryokbys
@usethedata -- I fixed  the bibliographic entry in JOSS branch. But I did not merge it to other branches as I do not know whether I need to do it..
