### 2020-11-08 #2768 @richardjgowers
I've started the review now.  I've raised a few issues on the software's repo about installation dependencies.  There's currently not even regression tests in place, so there's no way of verifying that what I've compiled is equivalent to the intended result (especially since it's noted that `-O3` optimisations can cause numerical errors...).  Ideally for each major feature there should be (at least) a regression test, maybe the existing examples can be repurposed here.
