### 2020-11-10 #2768 @ryokbys
@richardjgowers -- Thanks for your suggestions. As you mentioned, the examples are made for that purpose. But there was no test code or documentation about it. I added an installation documentation page in [nap documentation](http://ryokbys.web.nitech.ac.jp/contents/nap_docs/install.html) and wrote how to test at the bottom of the page.
