+++ pmd/mod_localflux.F90
@@ -1,6 +1,6 @@
 module localflux
 !
-!  Module for evaluation of local flux using some NEMD method.
+!  Module for evaluation of local flux using the color-charge NEMD.
 !
   use pmdio,only: csp2isp,nspmax
   use pmdmpi,only: nid2xyz
@@ -14,6 +14,8 @@ module localflux
 
   logical:: lflux = .false.  ! Flag to eval local flux.
   logical:: initialized = .false.
+  integer:: noutlflux = 1000
+  integer:: nskipout
   real(8),allocatable:: fluxg(:,:)  ! Global flux
   real(8),allocatable:: fluxl(:,:),fluxt(:,:)  ! Local flux
   integer:: nlx = 1  ! Num of divisions in a node for local regions
@@ -23,23 +25,32 @@ module localflux
   integer:: ng,ngx,ngy,ngz  ! Global num of divisions
   integer,allocatable:: idxl2g(:,:)  ! Conversion of index from local to global
   integer,allocatable:: nargn(:)  ! Num of atoms in local regions
-  real(8):: dlx,dly,dlz,dlxi,dlyi,dlzi
+  real(8):: dlx,dly,dlz,dlxi,dlyi,dlzi,evflux(3)
 
 contains
 !=======================================================================
-  subroutine init_lflux(myid,nx,ny,nz,lclrchg,mpi_world,iprint)
+  subroutine init_lflux(myid,nx,ny,nz,lclrchg,maxstp,mpi_world,iprint)
 !
 !  Initialize lflux.
 !
-    integer,intent(in):: myid,mpi_world,nx,ny,nz,iprint
+    use clrchg,only: clrfield
+    integer,intent(in):: myid,mpi_world,nx,ny,nz,iprint,maxstp
     logical,intent(in):: lclrchg
     
     integer:: inc,ix,iy,iz,mx,my,mz,ixyz,nxyz
     integer:: idxlfg,ilfgx,ilfgy,ilfgz
-    real(8):: anxi,anyi,anzi
+    real(8):: anxi,anyi,anzi,fext
     integer,allocatable:: idxl2gt(:)
     integer:: istat(mpi_status_size),itag,ierr
-    
+
+    if( noutlflux.eq.0 ) stop 'noutlflux must not be 0.'
+    nskipout = max(maxstp/noutlflux,1)
+
+!.....Set unit vector of external force on color charge
+    fext = sqrt(clrfield(1)**2 +clrfield(2)**2 +clrfield(3)**2)
+    evflux(1:3) = clrfield(1:3)/fext
+
+!.....Parallel setting
     anxi= 1d0/nx
     anyi= 1d0/ny
     anzi= 1d0/nz
@@ -114,14 +125,14 @@ contains
            ,ngx,ngy,ngz,ng
       print '(a,4f10.1)','   Normalized lengths(x,y,z) of local region = ' &
            ,dlx,dly,dlz
+      print '(a,3f7.3)', '   Unit vector along Fext = ',evflux(1:3)
 
       open(ionum,file=trim(cfoutlf),status='replace')
       write(ionum,'(a)') '#'
       write(ionum,'(a,4(1x,i0))') '# Local flux of regions, ngx,ngy,ngz,ng = '&
            ,ngx,ngy,ngz,ng
       write(ionum,'(a)') '# Order of data: idx=(igx-1)*ngy*ngz +(igy-1)*ngz +igz'
-      write(ionum,'(a)') '# Each line contains:  istp, time, ng*(jx,jy,jz)'
-      write(ionum,'(a)') '# Unit of flux: Ang/fs'
[…]
