+++ pmd/mod_pdens.F90
@@ -1,0 +1,202 @@
+module pdens
+!
+!  Module for evaluation of probability density.
+!
+  use pmdio,only: csp2isp,nspmax,get_vol
+  use pmdmpi,only: nid2xyz
+  implicit none
+  include 'mpif.h'
+  save
+
+  character(len=128):: paramsdir = '.'
+  character(len=128),parameter:: cfoutpd = 'out.pdens.cube'
+  integer,parameter:: ionum = 68
+
+  logical:: lpdens = .false.  ! Flag to eval local flux.
+  logical:: initialized = .false.
+  character(len=3):: cspc_pdens = 'non'
+  integer:: ispc_pdens
+  real(8),allocatable:: pdg(:)  ! Global pdens
+  real(8),allocatable:: pdl(:),pdt(:)  ! Local pdens
+  integer:: npx = 1  ! Num of divisions in a node for local regions
+  integer:: npy = 1  ! where flux is evaluated within.
+  integer:: npz = 1
+  integer:: nl
+  integer:: ng,ngx,ngy,ngz  ! Global num of divisions
+  integer,allocatable:: idxl2g(:,:)  ! Conversion of index from local to global
+  real(8):: dpx,dpy,dpz,dpxi,dpyi,dpzi
+  integer:: nacc
+
+contains
+!=======================================================================
+  subroutine init_pdens(myid,nx,ny,nz,hmat,specorder,mpi_world,iprint)
+!
+!  Initialize lflux.
+!
+    integer,intent(in):: myid,mpi_world,nx,ny,nz,iprint
+    character(len=3),intent(in):: specorder(nspmax)
+    real(8),intent(in):: hmat(3,3)
+    
+    integer:: inc,ix,iy,iz,mx,my,mz,ixyz,nxyz
+    integer:: idxg,ipx,ipy,ipz
+    real(8):: anxi,anyi,anzi,fext
+    integer,allocatable:: idxl2gt(:)
+    integer:: istat(mpi_status_size),itag,ierr
+
+    if( trim(cspc_pdens).eq.'non' ) then
+      stop 'ERROR: spcs_pdens must be specified.'
+    else
+      ispc_pdens = csp2isp(trim(cspc_pdens),specorder)
+    endif
+
+!.....Parallel setting
+    anxi= 1d0/nx
+    anyi= 1d0/ny
+    anzi= 1d0/nz
+    nxyz = nx*ny*nz
+
+!.....Number of local-flux regions in a node and global
+    nl = npx*npy*npz
+    ngx = npx*nx
+    ngy = npy*ny
+    ngz = npz*nz
+    ng = ngx*ngy*ngz
+
+    dpx = anxi /npx
+    dpy = anyi /npy
+    dpz = anzi /npz
+    dpxi = 1d0/dpx
+    dpyi = 1d0/dpy
+    dpzi = 1d0/dpz
+
+    allocate(pdl(nl),pdt(nl),idxl2g(nl,0:nxyz-1))
+    allocate(idxl2gt(nl))
+    pdl(:) = 0d0
+
+!.....Index conversion from local to global
+    call nid2xyz(myid,mx,my,mz)
+    inc = 0
+    do ix=1,npx
+      ipx = mx*npx +ix
+      do iy=1,npy
+        ipy = my*npy +iy
+        do iz=1,npz
+          ipz = mz*npz +iz
+          inc = inc +1
+          idxg = (ipx-1)*ngz*ngy +(ipy-1)*ngz +ipz
+          idxl2gt(inc) = idxg
+        enddo
+      enddo
+    enddo
+!.....Gather idxl2gt to make idxl2g in node-0
+    if( myid.eq.0 ) then
+      idxl2g(1:nl,0) = idxl2gt(1:nl)
[…]
