### 2020-11-19 a2937c0
fixed a bug in mod_pdens.
