### 2020-11-29 #2768 @richardjgowers
Just to check in, the regression tests are still outstanding and with the package known to have compilation quirks I think this is a show stopper in terms of ensuring reproducibility.
