### 2020-11-30 c99e93b
fixed bugs in fitpot and added makefile in fitpot_DNN_SiO.
