### 2020-11-30 8bcb47e
added test in examples/fp_*.
