+++ makefile
@@ -1,16 +1,25 @@
+.phony: all clean pmd fitpot test
+SHELL = /bin/bash
 
-.phony: all clean pmd fitpot FORCE
+all: pmd fitpot test
 
-all: pmd fitpot
+clean:
+	(cd pmd/ && make clean)
+	(cd fitpot/ && make clean)
 
-clean: FORCE
-	cd pmd; make clean
-	cd fitpot; make clean
+pmd:
+	(cd pmd/ && make pmd)
 
-pmd: FORCE
-	cd pmd; make pmd
+fitpot: pmd
+	(cd fitpot/ && make fitpot)
 
-fitpot: FORCE
-	cd fitpot; make fitpot
+test: pmd fitpot
+	@(cd pmd/ && make test)
+	@(cd fitpot/ && make test)
+	@echo " Run make test-fp after setting up PYTHONPATH for nappy."
+	@echo " To setup PYTHONPATH for nappy, see the documentation:"
+	@echo "   http://ryokbys.web.nitech.ac.jp/contents/nap_docs/install.html#setup-nappy-required-for-fppy"
 
-FORCE:
+test-fp:
+	@(cd examples/fp_LZP && make test)
+
