+++ makefile.in
@@ -1,34 +1,0 @@
-#!/usr/sh
-########################################################################
-# Makefile for parallel MD
-#     by Ryo Kobayashi
-########################################################################
-
-#-----------------------------------------------------------------------
-# to be generated by configure script
-#-----------------------------------------------------------------------
-CPP= @CPP@
-CPPFLAGS= @CPPFLAGS@
-#CPPFLAGS= -D__SHEAR__ -D__DISL__ -D__KOUKYU__ -D__WALL__
-MPIFC= @FC@
-MPIFLAGS= @FCFLAGS@
-
-all: pmd qmcl neb
-# 	cd pmd; make all
-# 	cp pmd/{pmd,nconv,combine,pmd2akr,akr2cna} .
-
-clean:
-	$(cd pmd; make clean)
-	$(cd mkconf; make clean)
-	$(cd qmcl; make clean)
-	$(cd neb; make clean)
-
-pmd: pmd/parallel_md.F
-	$(cd pmd; make pmd nconv combine pmd2akr akr2cna)
-
-qmcl: qmcl/qmcl.F
-	$(cd qmcl; make qmcl)
-
-neb: neb/neb.F
-	$(cd neb; make neb)
-
