### 2020-11-30 4c8d732
added makefile in examples/pmd_DNN_SiO
