### 2020-11-30 #2768 @ryokbys
Sorry for inconvenience. I just uploaded a new version v0.10.1 which includes test entries in makefile in root nap/ directory for the purpose of regression test. Please try the new one.
To test pmd and fitpot,
`$ cd nap/ ; make pmd fitpot ; make test`
To test fp.py,
`$ make test-fp`
which would take a few minutes.
