+++ pmd/pmd_core.F90
@@ -1,5 +1,5 @@
 !-----------------------------------------------------------------------
-!                     Last-modified: <2020-11-18 22:16:18 Ryo KOBAYASHI>
+!                     Last-modified: <2020-12-24 07:54:19 Ryo KOBAYASHI>
 !-----------------------------------------------------------------------
 ! Core subroutines/functions needed for pmd.
 !-----------------------------------------------------------------------
@@ -33,6 +33,7 @@ subroutine pmd_core(hunit,h,ntot0,tagtot,rtot,vtot,atot,stot &
   use clrchg,only: lclrchg,clrchg_force,rm_trans_clrchg,clr_init
   use localflux,only: lflux,accum_lflux
   use pdens,only: lpdens,accum_pdens
+  use time, only: sec2hms, accum_time
   implicit none
   include "mpif.h"
   include "./params_unit.h"
@@ -171,8 +172,10 @@ subroutine pmd_core(hunit,h,ntot0,tagtot,rtot,vtot,atot,stot &
     noutpmd = maxstp +1
   endif
 !.....perform space decomposition after reading atomic configuration
+  tmp = mpi_wtime()
   call space_decomp(hunit,h,ntot0,tagtot,rtot,vtot,chgtot,chitot,teitot &
        ,clrtot,myid_md,mpi_md_world,nx,ny,nz,nxyz,rc,rbuf,iprint)
+  call accum_time('space_decomp',mpi_wtime()-tmp)
 !.....Some conversions
   do i=1,natm
     ra(1:3,i)= ra(1:3,i) -sorg(1:3)
@@ -331,21 +334,21 @@ subroutine pmd_core(hunit,h,ntot0,tagtot,rtot,vtot,atot,stot &
 
   tcpu1= mpi_wtime()
   tcom = 0d0
-  tlspr = 0d0
-  tdump = 0d0
 
   call init_force(namax,natm,nsp,tag,chg,chi,myid_md,mpi_md_world, &
        iprint,h,rc,lvc,ifcoulomb,specorder,am)
 !-----copy RA of boundary atoms
   call check_size_and_parallel(sgm,vol,rc,anxi,anyi,anzi &
        ,nx,ny,nz,myid_md)
+  tmp = mpi_wtime()
   call bacopy(rc,myid_md,mpi_md_world,iprint,ifcoulomb &
        ,.true.,boundary)
+  call accum_time('ba_xxx',mpi_wtime()-tmp)
 !-----Make pair list
   tmp = mpi_wtime()
   call mk_lspr_para(namax,natm,nbmax,nb,nnmax,tag,ra,rc+rbuf,rc1nn &
        ,h,hi,anxi,anyi,anzi,lspr,ls1nn,iprint,.true.)
-  tlspr = tlspr +(mpi_wtime() -tmp)
+  call accum_time('lspr',mpi_wtime()-tmp)
 
 !.....Calc forces
   lstrs = lstrs0 .or. &
@@ -354,10 +357,12 @@ subroutine pmd_core(hunit,h,ntot0,tagtot,rtot,vtot,atot,stot &
        trim(cpctl).eq.'vv-Berendsen')
 !.....Cell is new at the first call of get_force
   lcell_updated = .true.
+  tmp = mpi_wtime()
   call get_force(namax,natm,tag,ra,nnmax,aa,strs,chg,chi,tei,stnsr &
        ,h,hi,tcom,nb,nbmax,lsb,lsex,nex,lsrc,myparity,nn,sv,rc &
[…]
