+++ pmd/mod_pairlist.F90
@@ -1,0 +1,498 @@
+module pairlist
+!-----------------------------------------------------------------------
+! Module for pair-list.
+!-----------------------------------------------------------------------
+  
+contains
+  subroutine init_pairlist(namax)
+    
+    
+  end subroutine init_pairlist
+!=======================================================================
+  subroutine mk_lspr_para(namax,natm,nbmax,nb,nnmax,tag,ra,rc,rc1nn &
+       ,h,hi,anxi,anyi,anzi,lspr,ls1nn,iprint,l1st)
+    implicit none
+    integer,intent(in):: namax,natm,nbmax,nb,nnmax,iprint
+    integer,intent(out):: lspr(0:nnmax,namax),ls1nn(0:nnmax,namax)
+    real(8),intent(in):: ra(3,namax),rc,rc1nn,anxi,anyi,anzi &
+         ,hi(3,3),h(3,3),tag(namax)
+    logical,intent(in):: l1st
+
+    integer:: i,j,k,l,m,n
+    integer:: mx,my,mz,kux,kuy,kuz,m1x,m1y,m1z,m1,ic,jc,ierr
+    real(8):: xi(3),xij(3),rij(3),rij2
+
+    integer,allocatable,save:: lscl(:),lshd(:)
+    real(8),save:: rc2,rcx,rcy,rcz,rcxi,rcyi,rczi,rc1nn2
+    integer,save:: lcx,lcy,lcz,lcx2,lcy2,lcz2,lcyz2,lcxyz2
+
+    if( l1st ) then
+      rc2= rc**2
+      rc1nn2 = rc1nn**2
+!-----make a linked cell list, LSCL
+      lcx=anxi/dsqrt(hi(1,1)**2+hi(1,2)**2+hi(1,3)**2)/rc
+      lcy=anyi/dsqrt(hi(2,1)**2+hi(2,2)**2+hi(2,3)**2)/rc
+      lcz=anzi/dsqrt(hi(3,1)**2+hi(3,2)**2+hi(3,3)**2)/rc
+!.....In case that system is thinner than rc, modify lc?.
+!.....but notice this modification does not correct results.
+      if( lcx.eq.0 ) lcx=1
+      if( lcy.eq.0 ) lcy=1
+      if( lcz.eq.0 ) lcz=1
+      lcx2= lcx +2
+      lcy2= lcy +2
+      lcz2= lcz +2
+      lcyz2=lcy2*lcz2
+      lcxyz2=lcx2*lcyz2
+      rcx= anxi/lcx
+      rcy= anyi/lcy
+      rcz= anzi/lcz
+      rcxi=1d0/rcx
+      rcyi=1d0/rcy
+      rczi=1d0/rcz
+!        write(6,'(a,3i8)') ' lcx,lcy,lcz=',lcx,lcy,lcz
+!        write(6,'(a,3i8)') ' lcx2,lcy2,lcz2=',lcx2,lcy2,lcz2
+!        write(6,'(a,i8)') ' lcxyz2=',lcxyz2
+!        write(6,'(a,3es12.4)') ' rcx,rcy,rcz=',rcx,rcy,rcz
+!-----allocate LSCL & LSHD after obtaining lcxyz2
+      if( allocated(lscl) ) deallocate(lscl,lshd)
+      allocate(lscl(namax+nbmax),lshd(lcxyz2))
+    endif
+
+!-----reset pair list, LSPR
+    lspr(0,:)= 0
+    ls1nn(0,:)= 0
+
+!-----reset headers
+    lshd(1:lcxyz2)= 0
+
+
+!-----construct a linked-cell list, LSCL, & a header list, LSHD
+    do i=1,natm+nb
+!-------assign a vector cell index
[…]
