+++ pmd/makefile.in
@@ -25,16 +25,17 @@ MPIFLAGS= @FCFLAGS@
 .F90.o: 
 	${MPIFC} -c ${MPIFLAGS} ${CPPFLAGS} -o $@ $<
 
-pmd= read_input.o util_vec.o util_rnd.o lasubs.o mk_lspr.o \
+pmd= read_input.o util_vec.o util_rnd.o lasubs.o \
 	common_parallel.o sort.o pmd_main.o pmd_core.o 
-pmc= prec_mc.o util_rnd.o mk_lspr.o read_input.o pmd_core.o util_vec.o \
+pmc= prec_mc.o util_rnd.o read_input.o pmd_core.o util_vec.o \
 	lasubs.o common_parallel.o sort.o
 comb= node_conv.o util_vec.o
 
 mods= mod_pmdvars.o mod_pmdio.o mod_pmdmpi.o zload.o mod_version.o mod_force.o \
 	mod_ttm.o descriptor.o metadynamics.o mod_constraints.o mod_vector.o \
 	mod_rdcfrc.o mod_structure.o mod_deform.o mod_util.o mod_element.o \
-	mod_extforce.o mod_colorcharge.o mod_localflux.o mod_pdens.o mod_time.o
+	mod_extforce.o mod_colorcharge.o mod_localflux.o mod_pdens.o mod_time.o \
+	mod_pairlist.o
 args= iargc.o getarg.o
 
 force_common= force_common.o
@@ -53,7 +54,7 @@ params=params_Ramas_FeH.h params_RK_FeH.h params_RK_WHe.h params_Ito3_WHe.h \
 	params_Mishin_Al.h params_Mishin_Ni.h params_AFS_W.h params_SC_Fe.h \
 	params_SM_Al.h
 
-clstr= mod_pmdio.o mod_util.o util_vec.o mk_lspr.o cluster.o
+clstr= mod_pmdio.o mod_util.o util_vec.o mod_pairlist.o cluster.o
 
 #-----------------------------------------------------------------------
 # Post process programs

[…]
