### 2020-12-25 4713a77
implemented linked cell reordering, but it causes bugs when using charges and speed improvement is not so great.
