### 2020-12-25 0fed2a3
made it enable to use linked-cell reordering if specified in input file.
