+++ pmd/read_input.F90
@@ -113,6 +113,9 @@ subroutine set_variable(ionum,cname)
   elseif( trim(cname).eq.'cutoff_buffer' ) then
     call read_r1(ionum,rbuf)
     return
+  elseif( trim(cname).eq.'reorder_arrays' ) then
+    call read_l1(ionum,lreorder)
+    return
   elseif( trim(cname).eq.'flag_damping' ) then
     call read_i1(ionum,ifdmp)
     return
