### 2020-12-26 9aa1f55
fixed bugs related to linked-cell reordering with BVS potential.
