+++ pmd/force_Morse.F90
@@ -1,6 +1,6 @@
 module Morse
 !-----------------------------------------------------------------------
-!                     Last modified: <2020-10-23 17:13:08 Ryo KOBAYASHI>
+!                     Last modified: <2020-12-26 00:39:35 Ryo KOBAYASHI>
 !-----------------------------------------------------------------------
 !  Parallel implementation of Morse pontential.
 !    - For BVS, see Adams & Rao, Phys. Status Solidi A 208, No.8 (2011)
