+++ pmd/force_Coulomb.F90
@@ -1,6 +1,6 @@
 module Coulomb
 !-----------------------------------------------------------------------
-!                     Last modified: <2020-10-23 17:16:55 Ryo KOBAYASHI>
+!                     Last modified: <2020-12-27 00:16:44 Ryo KOBAYASHI>
 !-----------------------------------------------------------------------
 !  Parallel implementation of Coulomb potential
 !  ifcoulomb == 1: screened Coulomb potential
@@ -1544,8 +1544,8 @@ contains
 
   end subroutine force_direct_cut
 !=======================================================================
-  subroutine force_screened_cut(namax,natm,tag,ra,nnmax,aa,strsl,chg,h,hi &
-       ,lspr,epi,esrl,iprint,lstrs,rc,l1st)
+  subroutine force_screened_cut(namax,natm,tag,ra,nnmax,aa,strsl &
+       ,chg,h,hi,lspr,epi,esrl,iprint,lstrs,rc,l1st)
 !
 !  Screened Coulomb with cutoff that uses rho_bvs.
 !  smoothing using vrc and dVdrc where
@@ -1553,6 +1553,7 @@ contains
 !
 !!$    use ZBL,only: zeta,dzeta
 !!$    use ZBL,only: interact_zbl => interact
+    use util,only: itotOf
     implicit none
     include "mpif.h"
     include "./params_unit.h"
@@ -1566,6 +1567,7 @@ contains
     logical,intent(in):: lstrs,l1st
 
     integer:: i,j,k,l,m,n,ierr,is,js,ixyz,jxyz,nconnect(4)
+    integer:: itot,jtot
     real(8):: xi(3),xj(3),xij(3),rij(3),dij,diji,dedr &
          ,dxdi(3),dxdj(3),x,y,z,epotl,epott,at(3),tmp &
          ,qi,qj,radi,radj,rhoij,terfc,texp &
@@ -1599,7 +1601,7 @@ contains
       do k=1,lspr(0,i)
         j=lspr(k,i)
         if(j.eq.0) exit
-        if(j.le.i) cycle
+!!$        if(j.le.i) cycle
         js= int(tag(j))
         if( .not.interact(is,js) ) cycle
         qj= chg(j)
@@ -1620,25 +1622,27 @@ contains
         dedr= -acc *qi*qj*diji *(1d0*diji*terfc +2d0/rhoij *sqpi *texp) -dvdrc
         tmp= 0.5d0 *( acc *qi*qj*diji *terfc -vrc -dvdrc*(dij-rc) )
 !.....potential
-        if( j.le.natm ) then
-          epi(i)= epi(i) +tmp
-          epi(j)= epi(j) +tmp
-          esrl = esrl +tmp +tmp
-        else
-          epi(i)= epi(i) +tmp
-          esrl = esrl +tmp
-        endif
+!!$        if( j.le.natm ) then
+!!$          epi(i)= epi(i) +tmp
+!!$          epi(j)= epi(j) +tmp
+!!$          esrl = esrl +tmp +tmp
+!!$        else
+!!$          epi(i)= epi(i) +tmp
+!!$          esrl = esrl +tmp
+!!$        endif
+        epi(i)= epi(i) +tmp
+        esrl = esrl +tmp
 !.....force
         aa(1:3,i)= aa(1:3,i) -dxdi(1:3)*dedr
-        aa(1:3,j)= aa(1:3,j) -dxdj(1:3)*dedr
+!!$        aa(1:3,j)= aa(1:3,j) -dxdj(1:3)*dedr
 !.....stress
         if( lstrs ) then
[…]
