### 2020-12-29 e8f8cfa
modified pairlist routine so that it does not use goto statements.
