+++ pmd/mod_pairlist.F90
@@ -5,7 +5,9 @@ module pairlist
   use force,only: luse_charge, luse_elec_temp
   use clrchg,only: lclrchg
   use pmdvars,only: chg,chi,tei,clr
+  use time,only: accum_time
   implicit none
+  include 'mpif.h'
   save
   
   integer,allocatable:: lscl(:),lshd(:)
@@ -13,6 +15,12 @@ module pairlist
   integer:: lcx,lcy,lcz,lcxyz,lcyz,lcx2,lcy2,lcz2,lcyz2,lcxyz2
   real(8),allocatable:: tmparr(:)
   integer:: ndmax
+
+!.....Arrays for Gonnet's algorithm
+  real(8),allocatable:: dlist(:)
+  integer,allocatable:: ilist(:),ileft(:),ia2ic(:)
+  real(8):: dh(3,3),vecc(3,26)
+  integer:: idcell(-1:1,-1:1,-1:1)
   
 contains
 !=======================================================================
@@ -124,7 +132,7 @@ contains
     logical,intent(in):: l1st
 
     integer:: i,j,k,l,m,n
-    integer:: mx,my,mz,kux,kuy,kuz,m1x,m1y,m1z,m1,ic,jc,ierr
+    integer:: mx,my,mz,kux,kuy,kuz,m1x,m1y,m1z,m1,ic,jc,ierr,mmax
     real(8):: xi(3),xij(3),rij(3),rij2
 
     call mk_lscl_para(namax,natm,nbmax,nb,ra,anxi,anyi,anzi,rc,rc1nn &
@@ -140,7 +148,7 @@ contains
       do my=0,lcy+1
         do mx=0,lcx+1
           m= mx*lcyz2 +my*lcz2 +mz +1
-          if(lshd(m).eq.0) goto 5
+          if(lshd(m).eq.0) cycle
           do kuz= -1,1
             m1z= mz +kuz
             if( m1z.lt.0 .or. m1z.gt.lcz+1 ) cycle
@@ -151,92 +159,301 @@ contains
                 m1x= mx +kux
                 if( m1x.lt.0 .or. m1x.gt.lcx+1 ) cycle
                 m1=m1x*lcyz2 +m1y*lcz2 +m1z +1
-                if(lshd(m1).eq.0) goto 6
+                if(lshd(m1).eq.0) cycle
 
                 i=lshd(m)
-1               continue
-!              if (natm.lt.i) goto 4
+                do while( i.gt.0 )
 
-                ic= int(tag(i))
-                xi(1:3)= ra(1:3,i)
+                  ic= int(tag(i))
+                  xi(1:3)= ra(1:3,i)
 
-                j=lshd(m1)
+                  j=lshd(m1)
+                  do while( j.gt.0 )
+                    if( j.le.i ) then
+                      j = lscl(j)
+                      cycle
+                    endif
 
-2               continue
-                if( j.le.i ) goto 3
-!          if (j.eq.i) goto 3
+                    jc= int(tag(j))
+                    xij(1:3)= ra(1:3,j) -xi(1:3)
+                    rij(1)= h(1,1)*xij(1) +h(1,2)*xij(2) +h(1,3)*xij(3)
+                    rij(2)= h(2,1)*xij(1) +h(2,2)*xij(2) +h(2,3)*xij(3)
+                    rij(3)= h(3,1)*xij(1) +h(3,2)*xij(2) +h(3,3)*xij(3)
+                    rij2= rij(1)**2 +rij(2)**2 +rij(3)**2
 
-                jc= int(tag(j))
-                xij(1:3)= ra(1:3,j) -xi(1:3)
[…]
