### 2020-12-29 7844c3e
modified error trap point in pairlist.
