### 2021-01-06 6feeb62
nproc options is now (finally) available in fp.py.
