### 2021-01-08 #2768 @foeroyingur
Sorry for the long delay! I've finally started the review. I created a couple of issues related to build instructions and test for parallel execution of pmd.
