### 2021-01-11 #2768 @foeroyingur
The issues I had have been fixed in v0.10.2 so I'm satsified. There are two points that could be improved but are acceptable as far as I can understand. One is the installation where there is a list of dependencies but no package management file. The other is manual testing which should preferably be automated.
