### 2021-01-12 #2768 @ryokbys
@foeroyingur Thank you for your valuable comment.
Since I am not familiar with package managements, I do not know how to create something like `yum` or `brew` repository(?) and currently not thinking of it.
For the testing, I have made `test` and `test-fp` entries in the makefile at nap root directory and I thought it is sufficiently automated. I'd appreciate any suggestion about further (or better) ways of automation.
