### 2021-01-12 #2768 @foeroyingur
Regarding package management, I was mainly thinking about the Python package dependencies which you can put in a requirements.txt. They can then be easily installed by `pip install -r requirements.txt`. It would also allow you to specify any version requirements. See e.g. https://pip.pypa.io/en/stable/reference/pip_install/#requirements-file-format.

You can setup automated testing, e.g., using [GitHub Actions](https://docs.github.com/en/free-pro-team@latest/actions), so that the tests are run in every commit.
