### 2021-01-12 #2768 @ryokbys
Preparing `requirements.txt` for installing python packages would not be so difficult and will be added soon.
On the other hand, I need some time for researching on GitHub Actions.
Thank you for your information.
