### 2021-01-15 c2cdc73
separated ODE solver and model in mod_ttm.
