+++ pmd/mod_ttm.F90
@@ -1,6 +1,6 @@
 module ttm
 !-----------------------------------------------------------------------
-!                     Last-modified: <2019-11-06 17:07:13 Ryo KOBAYASHI>
+!                     Last-modified: <2021-01-15 11:52:29 Ryo KOBAYASHI>
 !-----------------------------------------------------------------------
 !
 ! Module for two-temperature method (TTM).
@@ -31,9 +31,10 @@ module ttm
   integer:: nx,ny,nz,nxyz
 !.....Mesh size in reduced unit [0:1)
   real(8):: dx,dy,dz,area,darea,dr2
+!.....ODE solver; 1) Eular, 2) RK4 (4th order Runge-Kutta)
+  character(len=20):: csolver = 'Eular'
 !.....Time step in fs == dt in MD by default
   real(8):: dt_inner = -1.0
-  real(8):: dt
 !.....Number of inner loop in TTM
   integer:: nstp_inner = 1
 !.....Volume and area per mesh cell Ang^3 == dx*dy*dz
@@ -49,8 +50,9 @@ module ttm
   real(8):: alpha_max = -1d0
   real(8):: kappa_max, cete_min
 
-!.....Ce dependence on Te: none, polynomial, tanh, linear
+!.....Ce dependence on Te: none(0), polynomial(1), tanh(2), linear(3)
   character(len=128):: Ce_Tdep = 'none'
+  integer:: iCe_Tdep = 0
   real(8):: rho_e = 0.005
   real(8):: d_e = 20d0
   real(8):: c_0 = 1e-4
@@ -71,11 +73,13 @@ module ttm
 !.....Initial Te distribution: exp, homo, or read (from cTe_init)
   character(len=128):: cTe_init = 'homo'
   real(8):: Te_init = 300d0
-!.....T-dependence of pulse shape: stepwise or gaussian
+!.....T-dependence of pulse shape: stepwise(1, default) or gaussian(2)
   character(len=128):: ctype_pulse = 'stepwise'
-!.....Function shape of thermal diffusivity, kappa: DCrho or B2
+  integer:: itype_pulse = 1
+!.....Function shape of thermal diffusivity, kappa: DCrho(1, default) or B2(2)
 ! B2: from Eq.(B2) in PRB68,064114 (2003)
   character(len=128):: ctype_kappa = 'DCrho'
+  integer:: itype_kappa = 1
 !.....Prefactor in case of kappa_type = B2, in eV/(fs*Ang*K)
   real(8):: kappa0 = 6.2422d-7
 
@@ -107,8 +111,9 @@ module ttm
 !.....Atom to cell correspondance
   integer,allocatable:: a2c(:)
 
-!.....Type of coupling constant: constant_gp or constang_gmmp
+!.....Type of coupling constant: constant_gmmp(1) or constant_gp(2)
   character(len=128):: ctype_coupling = 'constant_gmmp'
+  integer:: itype_coupling = 1
 !.....e-ph coupling constant, g, in case of constant gp in eV/(fs*A^3*K)
 !.....Parameter for Ni from Zhigilei et al., J.Phys.Chem. C 113 (2009)
   real(8):: e_ph_const = 2.247e-9
@@ -168,10 +173,12 @@ contains
 !.....Convert fluence to intensity in unit of energy not temperature, here.
 !.....This is going to be converted to temperature by dividing by (rho*Ce)
[…]
