### 2021-01-17 #2768 @usethedata
I have verified that all of the DOIs resolve correctly.

@ryokbys -- We're just about there.  One nit I suggest fixing. In line 188 of paper.bib `s/van der Waals Corrections/van der {W}aals Corrections/`. It's a minor issue to get a name capitalization right.  You got all of the other forced upper cases escaped through.

When you've got that, make an archive at Zenodo and post the DOI for the archive here.
