+++ JOSS_paper/paper.bib
@@ -184,8 +184,8 @@
 }
 
 @ARTICLE{Morawietz2013-qq,
-  title    = "A {Density-Functional} {Theory-Based} Neural Network Potential
-              for Water Clusters Including van der Waals Corrections",
+  title    = "A Density-Functional Theory-Based Neural Network Potential
+              for Water Clusters Including van der {W}aals Corrections",
   author   = "Morawietz, Tobias and Behler, J{\"o}rg",
   journal  = "J. Phys. Chem. A",
   volume   =  117,
