### 2021-01-18 #2768 @ryokbys
I fixed the capitalization issue in paper.bib and the JOSS branch was merged to the master branch.
I have a Zenode DOI for nap and it seems automatically updated to point the newest version of it, which is now 10.5281/zenodo.4429382. But I don't know how to reflect the updated DOI to generated PDF. (I tried and made a newer release v0.10.3 after merging JOSS branch..., it did not work.) Please let me know what to do if I should do something.
