### 2021-01-19 bd86551
corrected typos in JOSS/paper.md.
