+++ JOSS_paper/paper.md
@@ -21,9 +21,9 @@ bibliography: paper.bib
 
 # Summary
 
-The **nap** is a package for molecular dynamics (MD) simulation consisting of an MD program (**pmd**) than can perform large-scale simulation using a spatial-decomposition technique and two parameter-optimization programs: one for classical (CL) potentials (**fp.py**) and another for machine-learning (ML) potentials (**fitpot**).
+The **nap** is a package for molecular dynamics (MD) simulation consisting of an MD program (**pmd**) that can perform large-scale simulation using a spatial-decomposition technique and two parameter-optimization programs: one for classical (CL) potentials (**fp.py**) and another for machine-learning (ML) potentials (**fitpot**).
 Since the numbers of parameters to be optimized are much different between CL and ML potentials, optimization approaches for them are also different; meta-heuristic global minimum-search algorithms for the CL potentials, in which the numbers of parameters are usually much less than one hundred, and gradient-based methods for the ML potentials.
-The parameters of CL potentials can be optimized to any target quantity that can be computed using the potentials since meta-heuristic methods do not require the derivatives of the quantity with respect to parameters. On the other hand, ML-potential parameters can be optimized to only energies, forces on atoms and stress components of reference systems, mainly because gradient-based methods require derivatives of other quantities with respect to parameters, and the analytical derivatives and the coding of them are usually painful and sometimes impossible.
+The parameters of CL potentials can be optimized to any target quantity that can be computed using the potentials since meta-heuristic methods do not require the derivatives of the quantity with respect to parameters. On the other hand, ML-potential parameters can be optimized to only energies, forces on atoms and stress components of reference systems, mainly because gradient-based methods require the derivatives of other quantities with respect to parameters, and the analytical derivatives and the coding of them are usually painful and sometimes impossible.
 Potentials can be used in combination with any other potential, such as pair and angular potentials, short-range and long-range potentials, CL and ML potentials.
 With using the **nap** package, users can perform MD simulation of solid-state materials with the choice of different levels of complexity (CL or ML) by creating interatomic potentials optimized to quantum-mechanical calculation data even if no potential is available.
 
