### 2021-01-19 #2768 @ryokbys
I found some typos and corrected them on master branch.
And the DOI of this paper in the PDF seems wrong, but I am not sure it is really wrong or not (I could think the link works after it is published...).
