### 2021-01-19 #2768 @kthyng
Hi @ryokbys! I'll be taking over to get your paper through the rest of the process.

Can you update the metadata for your Zenodo archive so that the title and author list match your JOSS paper exactly?
