### 2021-01-19 #2768 @kthyng
> I found some typos and corrected them on master branch.

Ok

> And the DOI of this paper in the PDF seems wrong, but I am not sure it is really wrong or not (I could think the link works after it is published...).

The DOI in the pdf is what will be assigned when it is submitted at the end of this process.
