### 2021-01-20 a672510
modified affiliation in JOSS paper.md
